added WIP for map validation. checking the player position is valid(no more than 1 and exists. checking the map width in order to know the boundaries.

+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: vovashko <vovashko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/11/18 15:33:40 by vovashko      #+#    #+#                 */
-/*   Updated: 2024/11/30 13:52:25 by vshkonda      ########   odam.nl         */
+/*   Updated: 2024/12/02 13:15:59 by vshkonda      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ typedef struct s_color
 typedef struct s_map_file_data
 {
 	char *file;
-	int width;
+	size_t width;
 	int height;
 	t_color *floor_color;
 	t_color *ceiling_color;
